chore: file updates

+++ src/pages/project/ProjectSummary.js
@@ -1,39 +1,67 @@
-import Avatar from "../../components/Avatar";
+import Avatar from '../../components/Avatar'
 import { useFirestore } from '../../hooks/useFirestore'
-import { useAuthContext } from "../../hooks/useAuthContext";
-import { useHistory } from "react-router-dom";
+import { useAuthContext } from '../../hooks/useAuthContext'
+import { useNavigate } from 'react-router-dom'
+import { useState } from 'react'
 
-export default function ProjectSummary({ project }) {
-	const { deleteDocument } = useFirestore('projects')
+export default function ProjectSummary({ project, setIsEditing }) {
+	const { deleteDocument, updateDocument } = useFirestore('projects')
 	const { user } = useAuthContext()
-	const history = useHistory()
+	const navigate = useNavigate()
+	const [isHovered, setIsHovered] = useState(null)
 
-	const handleClick = (e) => {
+	const archiveProject = () => {
+		const confirmation = window.confirm('Are you sure you want to archive this project?')
+		if (!confirmation) return
+		updateDocument(project.id, { status: 'archived' })
+		navigate('/')
+	}
+	const deleteProject = () => {
+		const confirmation = window.confirm('Are you sure you want to delete this project?')
+		if (!confirmation) return
 		deleteDocument(project.id)
-		history.push('/')
+		navigate('/')
+	}
+
+	const removeAssignedUser = (id) => {
+		if (project.assignedUsersList.length < 2) {
+			alert('You must have at least 1 assigned user')
+			return
+		}
+		const usersToKeep = project.assignedUsersList.filter(user => user.id !== id)
+		console.log(usersToKeep)
+		updateDocument(project.id, { assignedUsersList: usersToKeep })
+		alert('user has been removed from project')
 	}
 
 	return (
 		<div>
-			<div className="project-summary">
-				<h2 className="page-title">{project.name}</h2>
-				<p>By {project.createdBy.displayName}</p>
-				<p className="due-date">
-					Project due by {project.dueDate.toDate().toDateString()}
-				</p>
-				<p className="details">
-					{project.details}
-				</p>
-				<h4>Project is assigned to:</h4>
+			<div className={`project-summary ${project.status.toLowerCase()}`}>
+				<div className="status">{project.status.toUpperCase()}</div>
+				<p className="maestro"><a href={`https://maestro.alight.com/maestro/editIssue.html?issueId=${project.maestro}`} title={`Open Maestor ticket #${project.maestro}`} target="_blank" rel="noreferrer noopener">{project.maestro}</a></p>
+				<h2 className="client">{project.client}</h2>
+				<p className="details">{project.details}</p>
+				<p className="billing">{project.billing}</p>
+				<p className="date devstart">{project.devStartDate.toDate().toDateString()}</p>
+				<p className="date devqa">{project.devQADate.toDate().toDateString()}</p>
+				<p className="date qa">{project.qaDate.toDate().toDateString()}</p>
+				<p className="date qc">{project.qcDate.toDate().toDateString()}</p>
+				<p className="date prod">{project.prodDate.toDate().toDateString()}</p>
+				<p className="created-by">Created by {project.createdBy.displayName}</p>
+				<h4>Assigned to:</h4>
 				<div className="assigned-users">
 					{project.assignedUsersList.map(user => (
-						<div key={user.id}>
+						<div key={user.id} className="wrapper" onMouseEnter={() => setIsHovered(user.id)} onMouseLeave={() => setIsHovered(null)}>
+							{isHovered === user.id && <span className="remove-user" onClick={() => removeAssignedUser(user.id)}></span>}
 							<Avatar src={user} />
+							<span className="display-name" style={{ fontSize: '.675rem'}}>{user.displayName}</span>
 						</div>
 					))}
 				</div>
 			</div>
-			{user.uid === project.createdBy.id && <button className="btn" onClick={handleClick}>Mark as Complete</button>}
+			{project.status !== 'archived' && <button className="btn" onClick={() => setIsEditing(true)}>Edit Project</button>}
+			<button className="btn" style={project.status !== 'archived' ? {marginLeft: '1rem'} : {}} onClick={deleteProject}>Delete Project</button>
+			{project.status !== 'archived' && <button className="btn" style={{marginLeft: '1rem'}} onClick={archiveProject}>Archive Project</button>}
 		</div>
 	)
 }
